Extract MultiParentCasperBondingSpec
